src/requirements.txt
# The row below allow docker-compose to build images even if 
# geonode-mapstore-client-3.3.0 is a broken dependency. Can be deleted once geonode-mapstore-client
# 3.3.0 will be available 
django-countries==7.3.2
django-crispy-forms==1.14.0
django-select2==7.10.0
django-geojson==3.2.1
pandas==1.4.4
httpx==0.23.0
-e git+https://github.com/GeoNode/geonode-mapstore-client.git@3.3.x#egg=django_geonode_mapstore_client
-e git+https://github.com/GeoNode/geonode.git@3.3.x#egg=GeoNode
# GeoNode==3.3.2.post2
